Update name_station+anom for Ox
reverse order and remove coverage check from name station.
It now uses the anom table to get its names and stations, so Ox is captured, and the coverage check was done on No2 and O3 before building anom. This saves us updating the coverage table with Ox, but might decide to do this down the line